backend/requirements.txt
requests==2.32.4
Flask==3.1.1
flask-cors==6.0.1
python-dotenv==1.1.1
